data-capture/data-capture-streaming-initiator/requirements.txt
# Function dependencies, for example:
# package>=version
firebase_admin
google-cloud-firestore
google-cloud-pubsub